requirements.txt
streamlit
yt-dlp
tk
openai
python-docx
python-dotenv